feat: Add customer requests view and data type
Introduces a new data type `CustomerRequest` to store and manage customer feedback or feature requests.
Adds a corresponding view `CustomerRequestsView` to the application, allowing users to interact with these requests.
Updates the `ProjectData` interface to include an optional array of `customerRequests`.
Adds a new navigation item in the `Header` component for the `Requests` view.